move ConnRequest into dedicated file
